+++ public/js/fleets.js
@@ -5,7 +5,6 @@ $(document).ready(function(){
         var optionSelected = $("option:selected", this);
         var valueSelected = this.value; 
         //console.log(valueSelected);
-
         getPlane(valueSelected);
     })
 });
@@ -18,21 +17,5 @@ function getPlane(planeId) {
                 continue;
             $('#'+key).val(data[key]); 
         }
-
-        /*
-        $('#manufacturer').val(data.manufacturer);
-        $('#model').val(data.model);
-        $('#price').val(data.price);
-        $('#seats').val(data.seats);
-        $('#reach').val(data.reach);
-        $('#cruise').val(data.cruise);
-        $('#takeoff').val(data.takeoff);
-        $('#hourly').val(data.hourly);
-        */
     });
-    /*
-    $.get(url, function(data, status){
-        console.log(data);
-    });
-    */
 }